// arkalis/har-types.d.ts
export type NetlogEventParams = {
  REQUEST_ALIVE: undefined                      // type 2
  URL_REQUEST_START_JOB: {                      // type 106
    url: string
    method: "GET" | "POST" | "PUT" | "DELETE" | "HEAD" | "OPTIONS" | "PATCH"
    initiator: string | "not an origin"
    load_flags: number
    network_isolation_key: string
    request_type: "main frame" | "sub frame" | "other"
    priority?: number
    site_for_cookies: string
    traffic_annotation?: number
    upload_id?: string
    net_error?: number      // only when there's an error
  }
  HTTP_TRANSACTION_SEND_REQUEST_HEADERS: {      // type 171
    headers: string[]       // "Host: google.com", "Connection: keep-alive", ...
    line: string            // "GET / HTTP/1.1\r\n"
  },
  HTTP_TRANSACTION_HTTP2_SEND_REQUEST_HEADERS: {
    headers: string[]       // "Host: google.com", "Connection: keep-alive", ...
  },
  HTTP2_STREAM_UPDATE_SEND_WINDOW: {
    delta: number
    stream_id: number
    window_size: number
  },
  SSL_SOCKET_BYTES_SENT: {
    byte_count: number
    bytes: string   // base64 encoded
  },
  HTTP_TRANSACTION_READ_RESPONSE_HEADERS: {     // type 176
    headers: string[]       // "HTTP/1.1 200 OK", "Server: nginx", ...
  },
  HTTP_TRANSACTION_READ_BODY: undefined,
  URL_REQUEST_JOB_FILTERED_BYTES_READ: {
    byte_count: number
    bytes: string           // base64 encoded
  },
  HTTP_CACHE_READ_DATA: undefined,
  COOKIE_INCLUSION_STATUS: {
    status: string
    name: string
    domain: string
    path: string
    operation: "send" | "store"
  }
}

export interface NetlogJson {
  constants: Constants
  events: NetlogEvent<undefined>[]
}

interface Constants {
  activeFieldTrialGroups: string[]
  addressFamily: AddressFamily
  certPathBuilderDigestPolicy: CertPathBuilderDigestPolicy
  certStatusFlag: CertStatusFlag
  certVerifierFlags: CertVerifierFlags
  clientInfo: ClientInfo
  dnsQueryType: DnsQueryType
  loadFlag: LoadFlag
  loadState: LoadState
  logCaptureMode: string
  logEventPhase: LogEventPhase
  logEventTypes: LogEventTypes
  logFormatVersion: number
  logSourceType: LogSourceType
  netError: NetError
  quicError: QuicError
  quicRstStreamError: QuicRstStreamError
  secureDnsMode: SecureDnsMode
  timeTickOffset: number
}

interface AddressFamily {
  ADDRESS_FAMILY_IPV4: number
  ADDRESS_FAMILY_IPV6: number
  ADDRESS_FAMILY_UNSPECIFIED: number
}

interface CertPathBuilderDigestPolicy {
  kStrong: number
  kWeakAllowSha1: number
}

interface CertStatusFlag {
  AUTHORITY_INVALID: number
  CERTIFICATE_TRANSPARENCY_REQUIRED: number
  COMMON_NAME_INVALID: number
  CT_COMPLIANCE_FAILED: number
  DATE_INVALID: number
  INVALID: number
  IS_EV: number
  KNOWN_INTERCEPTION_BLOCKED: number
  KNOWN_INTERCEPTION_DETECTED: number
  NAME_CONSTRAINT_VIOLATION: number
  NON_UNIQUE_NAME: number
  NO_REVOCATION_MECHANISM: number
  PINNED_KEY_MISSING: number
  REVOKED: number
  REV_CHECKING_ENABLED: number
  SHA1_SIGNATURE_PRESENT: number
  SYMANTEC_LEGACY: number
  UNABLE_TO_CHECK_REVOCATION: number
  VALIDITY_TOO_LONG: number
  WEAK_KEY: number
  WEAK_SIGNATURE_ALGORITHM: number
}

interface CertVerifierFlags {
  VERIFY_DISABLE_NETWORK_FETCHES: number
}

interface ClientInfo {
  cl: string
  command_line: string
  name: string
  official: string
  os_type: string
  version: string
  version_mod: string
}

interface DnsQueryType {
  A: number
  AAAA: number
  HTTPS: number
  PTR: number
  SRV: number
  TXT: number
  UNSPECIFIED: number
}

interface LoadFlag {
  BYPASS_CACHE: number
  BYPASS_PROXY: number
  CAN_USE_RESTRICTED_PREFETCH: number
  DISABLE_CACHE: number
  DISABLE_CERT_NETWORK_FETCHES: number
  DISABLE_CONNECTION_MIGRATION_TO_CELLULAR: number
  DO_NOT_SAVE_COOKIES: number
  DO_NOT_USE_EMBEDDED_IDENTITY: number
  IGNORE_LIMITS: number
  MAIN_FRAME_DEPRECATED: number
  NORMAL: number
  ONLY_FROM_CACHE: number
  PREFETCH: number
  RESTRICTED_PREFETCH: number
  SKIP_CACHE_VALIDATION: number
  SKIP_VARY_CHECK: number
  SUPPORT_ASYNC_REVALIDATION: number
  VALIDATE_CACHE: number
}

interface LoadState {
  CONNECTING: number
  DOWNLOADING_PAC_FILE: number
  ESTABLISHING_PROXY_TUNNEL: number
  IDLE: number
  OBSOLETE_WAITING_FOR_APPCACHE: number
  READING_RESPONSE: number
  RESOLVING_HOST: number
  RESOLVING_HOST_IN_PAC_FILE: number
  RESOLVING_PROXY_FOR_URL: number
  SENDING_REQUEST: number
  SSL_HANDSHAKE: number
  WAITING_FOR_AVAILABLE_SOCKET: number
  WAITING_FOR_CACHE: number
  WAITING_FOR_DELEGATE: number
  WAITING_FOR_RESPONSE: number
  WAITING_FOR_STALLED_SOCKET_POOL: number
}

interface LogEventPhase {
  PHASE_BEGIN: number
  PHASE_END: number
  PHASE_NONE: number
}
// ...
interface LogSourceType {
  BIDIRECTIONAL_STREAM: number
  CERT_VERIFIER_JOB: number
  CERT_VERIFIER_TASK: number
  COOKIE_STORE: number
  CT_TREE_STATE_TRACKER: number
  DISK_CACHE_ENTRY: number
  EXPONENTIAL_BACKOFF_THROTTLING: number
  HOST_CACHE_PERSISTENCE_MANAGER: number
  HOST_RESOLVER_IMPL_JOB: number
  HTTP2_SESSION: number
  HTTP3_SESSION: number
  HTTP_AUTH_CONTROLLER: number
  HTTP_PROXY_CONNECT_JOB: number
  HTTP_SERVER_PROPERTIES: number
  HTTP_STREAM_JOB: number
  HTTP_STREAM_JOB_CONTROLLER: number
  MEMORY_CACHE_ENTRY: number
  NETWORK_QUALITY_ESTIMATOR: number
  NETWORK_SERVICE_HOST_RESOLVER: number
  NONE: number
  PAC_FILE_DECIDER: number
  PROXY_CLIENT_SOCKET: number
  QUIC_CONNECTION_MIGRATION: number
  QUIC_SESSION: number
  QUIC_STREAM_FACTORY_JOB: number
  SERVER_PUSH_LOOKUP_TRANSACTION: number
  SOCKET: number
  SOCKS_CONNECT_JOB: number
  SSL_CONNECT_JOB: number
  TRANSPORT_CONNECT_JOB: number
  TRIAL_CERT_VERIFIER_JOB: number
  UDP_SOCKET: number
  URL_REQUEST: number
  WEB_TRANSPORT_CLIENT: number
}
// ...
interface QuicRstStreamError {
  QUIC_APPLICATION_DONE_WITH_STREAM: number
  QUIC_BAD_APPLICATION_PAYLOAD: number
  QUIC_DATA_AFTER_CLOSE_OFFSET: number
  QUIC_DUPLICATE_PROMISE_URL: number
  QUIC_ERROR_PROCESSING_STREAM: number
  QUIC_HEADERS_TOO_LARGE: number
  QUIC_INVALID_PROMISE_METHOD: number
  QUIC_INVALID_PROMISE_URL: number
  QUIC_MULTIPLE_TERMINATION_OFFSETS: number
  QUIC_PROMISE_VARY_MISMATCH: number
  QUIC_PUSH_STREAM_TIMED_OUT: number
  QUIC_REFUSED_STREAM: number
  QUIC_RST_ACKNOWLEDGEMENT: number
  QUIC_STREAM_CANCELLED: number
  QUIC_STREAM_CLOSED_CRITICAL_STREAM: number
  QUIC_STREAM_CONNECTION_ERROR: number
  QUIC_STREAM_CONNECT_ERROR: number
  QUIC_STREAM_DECODER_STREAM_ERROR: number
  QUIC_STREAM_DECOMPRESSION_FAILED: number
  QUIC_STREAM_ENCODER_STREAM_ERROR: number
  QUIC_STREAM_EXCESSIVE_LOAD: number
  QUIC_STREAM_FRAME_ERROR: number
  QUIC_STREAM_FRAME_UNEXPECTED: number
  QUIC_STREAM_GENERAL_PROTOCOL_ERROR: number
  QUIC_STREAM_ID_ERROR: number
  QUIC_STREAM_INTERNAL_ERROR: number
  QUIC_STREAM_MISSING_SETTINGS: number
  QUIC_STREAM_NO_ERROR: number
  QUIC_STREAM_PEER_GOING_AWAY: number
  QUIC_STREAM_REQUEST_INCOMPLETE: number
  QUIC_STREAM_REQUEST_REJECTED: number
  QUIC_STREAM_SETTINGS_ERROR: number
  QUIC_STREAM_STREAM_CREATION_ERROR: number
  QUIC_STREAM_TTL_EXPIRED: number
  QUIC_STREAM_UNKNOWN_APPLICATION_ERROR_CODE: number
  QUIC_STREAM_VERSION_FALLBACK: number
  QUIC_STREAM_WEBTRANSPORT_BUFFERED_STREAMS_LIMIT_EXCEEDED: number
  QUIC_STREAM_WEBTRANSPORT_SESSION_GONE: number
  QUIC_UNAUTHORIZED_PROMISE_URL: number
}

interface SecureDnsMode {
  Automatic: number
  Off: number
  Secure: number
}

// export enum NetlogPhase {
//   PHASE_NONE = 0,
//   PHASE_BEGIN = 1,
//   PHASE_END = 2
// }

export type NetlogPhase = number   // TODO: probably a better way to do this with an enum

export interface NetlogEvent<T> {
  params?: T
  phase: NetlogPhase
  source: NetlogSource
  time: string
  type: number
}
// ...
interface NetlogSource {
  id: number
  start_time: string
  type: number
}